- can now get image cover from scanning
