+++ code/app/src/main/res/drawable/background_gradient.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:startColor="#ffffff"
+        android:centerColor="#ffff"
+        android:endColor="#006499"
+        android:angle="90" />
+</shape>
